Include content within output

includes2() is a version of the includes() helper function from rmarkdown that uses yml_blank() instead of NULL as the argument defaults, as ymlthis treats NULLs as literal YAML syntax ("null").

Usage

includes2(
  in_header = yml_blank(),
  before_body = yml_blank(),
  after_body = yml_blank()
)

Arguments

in_header

One or more files with content to be included in the header of the document.

before_body

One or more files with content to be included before the document body.

after_body

One or more files with content to be included after the document body.

Value

a list

Examples

# \donttest{
yml() %>%
  yml_output(
    pdf_document(includes = includes2(after_body = "footer.tex"))
  )
#> ---
#> date: '`r format(Sys.Date())`'
#> output:
#>   pdf_document:
#>     includes:
#>       after_body: footer.tex
#> ---
#> 
# }