Generate a full YAML template for your pkgdown site

pkgdown includes three helpful pkgdown::template_*() functions to generate the navbar, reference, and article YAML for the _pkgdown.yml file. pkgdown_template() is a wrapper function that runs all three, combines them, and converts them to a yml object. You may also pass pkgdown::template_*() functions to as_yml() to convert the individual sections. pkgdown_template() is particularly useful with use_pkgdown_yml() to write directly to the _pkgdown.yml file.

pkgdown_template(path = ".")

Arguments

path
The path to your package directory

Value

a yml object

See also

Examples

if (FALSE) {
# requires this to be a package directory
pkgdown_template() %>%
  use_pkgdown_yml()
}