Top-level YAML options for pagedown

pagedown offers several output functions for paginated output, resumes, business cards, theses, and morem as described in the pagedown vignette. pagedown also accepts a few custom top-level YAML. See pagedown_business_card_template() for more on setting up the YAML for a business card.

yml_pagedown_opts(
  .yml,
  toc = yml_blank(),
  toc_title = yml_blank(),
  lot = yml_blank(),
  lot_title = yml_blank(),
  chapter_name = yml_blank(),
  links_to_footnotes = yml_blank(),
  paged_footnotes = yml_blank()
)

Arguments

.yml

a yml object created by yml(), as_yml(), or returned by a yml_*() function

toc

Logical. Use a table of contents?

toc_title

The title for the table of contents. Note that the actual YAML field is toc-title

lot

Logical. Use a list of figures?

lot_title

The title for the list of figures. Note that the actual YAML field is lot-title

chapter_name

The chapter title prefix

links_to_footnotes

Logical. Transform all the URLs to footnotes? Note that the actual YAML field is links-to-footnotes

paged_footnotes

Logical. Render notes as footnotes? Note that the actual YAML field is paged-footnotes

Value

a yml object

See also

Examples

yml() %>%
  yml_pagedown_opts(
    toc = TRUE,
    toc_title = "TOC",
    chapter_name = c("CHAPTER\\ ", "."),
    links_to_footnotes = TRUE
  )
#> ---
#> date: '`r format(Sys.Date())`'
#> toc: true
#> toc-title: TOC
#> chapter_name:
#> - 'CHAPTER\ '
#> - '.'
#> links-to-footnotes: true
#> ---
#>