Set YAML related to rticles output formats

The rticles package include numerous output formats specific to academic journals. All of these can take YAML similar to pdf_document(). Additionally, two templates include custom YAML, rticles::sage_article() and rticles::sim_article(). See the help pages for these functions for more details and the sources of the LaTeX templates used for each.

yml_rticles_opts(
  .yml,
  title = yml_blank(),
  runninghead = yml_blank(),
  author = yml_blank(),
  authormark = yml_blank(),
  address = yml_blank(),
  corrauth = yml_blank(),
  corres = yml_blank(),
  email = yml_blank(),
  abstract = yml_blank(),
  received = yml_blank(),
  revised = yml_blank(),
  accepted = yml_blank(),
  keywords = yml_blank(),
  bibliography = yml_blank(),
  longtable = yml_blank(),
  classoption = yml_blank(),
  header_includes = yml_blank(),
  include_after = yml_blank(),
  ...
)

rticles_author(name = yml_blank(), num = yml_blank())

rticles_address(name = yml_blank(), org = yml_blank())

rticles_corr_author(
  name = yml_blank(),
  author = yml_blank(),
  address = yml_blank()
)

Arguments

.yml
a yml object created by yml(), as_yml(), or returned by a yml_*() function

title
Title of the manuscript

runninghead
A character vector, a short author list for the header (sage_article)

author
A list of authors, containing name and num fields (sage_article, sim_article). Use rticles_author() or a list to specify.

authormark
A character vector, the short author list for the header (sim_article)

address
list containing num and org for defining author affiliations (sage_article, sim_article). Use rticles_address() or a list to specify.

corrauth
corresponding author name and address (sage_article). Use rticles_corr_author() or a list to specify.

corres
author and address for correspondence (sim_article). Use rticles_corr_author() or a list to specify.

email
The email of the correspondence author (sage_article)

abstract
The abstract, limited to 200 words (sage_article), 250 words (sim_article)

received, revised, accepted
The dates of submission, revision, and acceptance of the manuscript (sim_article)

keywords
The keywords for the article (sage_article), up to 6 keywords (sim_article)

bibliography
BibTeX .bib file name (sage_article, sim_article)

longtable
Logical. Include the longtable package? Used by default from pandoc to convert markdown to LaTeX code (sim_article)

classoption
a character vector of classoption options for the sagej class (sage_article)

header_includes
additional LaTeX code to include in the header, before the \\begin\{document\} statement (sage_article, sim_article). Note that the actual YAML field is header-includes

include_after
additional LaTeX code to include before the \\end\{document\} statement (sage_article, sim_article). Note that the actual YAML field is include-after.

...
additional named R objects, such as characters or lists, to transform into YAML

name
The author's name

num
The author's number or address number

org
The author's organization

Value

a yml object

See also

Examples

yml() %>%
  yml_rticles_opts(received = "09-12-2014")

#> ---
#> author: runner
#> date: '`r format(Sys.Date())`'
#> received: 09-12-2014
#> ---
#>